+++ packages/common/src/index.ts
@@ -2,6 +2,5 @@ export * from './Constants';
 export * from './Util';
 export * from './Types';
 export * from 'discord-api-types/v10';
-export * from './applyToClass';
 export * from './Collection';
 export * from './Logger';