+++ packages/helpers/src/commands/index.ts
@@ -1,3 +1,3 @@
+export * from "./contextMenu/ContextCommand";
 export * from "./slash/SlashCommand";
 export * from "./slash/SlashCommandOption";
-export * from "./contextMenu/ContextCommand";