+++ packages/helpers/src/components/ActionRow.ts
@@ -1,9 +1,9 @@
-import { APIActionRowComponent, APIMessageActionRowComponent, ComponentType, TypeArray } from '@biscuitland/common';
-import { MessageComponents, createComponent } from '../Utils';
-import { BaseComponent } from './BaseComponent';
+import { APIActionRowComponent, APIMessageActionRowComponent, ComponentType, TypeArray } from "@biscuitland/common";
+import { MessageComponents, createComponent } from "../Utils";
+import { BaseComponent } from "./BaseComponent";
 
 export class MessageActionRow<T extends MessageComponents> extends BaseComponent<APIActionRowComponent<APIMessageActionRowComponent>> {
-  constructor({ components, ...data }: Partial<APIActionRowComponent<APIMessageActionRowComponent>>) {
+  constructor({ components, ...data }: Partial<APIActionRowComponent<APIMessageActionRowComponent>> = {}) {
     super({ ...data, type: ComponentType.ActionRow });
     this.components = (components?.map(createComponent) ?? []) as T[];
   }
@@ -22,7 +22,7 @@ export class MessageActionRow<T extends MessageComponents> extends BaseComponent
   toJSON(): APIActionRowComponent<APIMessageActionRowComponent> {
     return {
       ...this.data,
-      components: this.components.map((c) => c.toJSON())
-    } as APIActionRowComponent<ReturnType<T['toJSON']>>;
+      components: this.components.map((c) => c.toJSON()),
+    } as APIActionRowComponent<ReturnType<T["toJSON"]>>;
   }
 }