+++ packages/ws/src/defaults.ts
@@ -1,12 +1,10 @@
 export const GatewayManagerDefaultOptions = {
-  compress: false,
   intents: 0,
   properties: {
     os: process.platform,
     browser: 'Biscuit',
     device: 'Biscuit'
   },
-  url: 'wss://gateway.discord.gg',
   version: 10,
   totalShards: 1,
   lastShardId: 0,