+++ packages/ws/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @biscuitland/ws
 
+## 3.0.9
+
+### Patch Changes
+
+- silly bug on options
+
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @biscuitland/common@0.0.8
+
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @biscuitland/common@0.0.7
+
+## 3.0.6
+
+### Patch Changes
+
+- I meessed up
+- Updated dependencies
+  - @biscuitland/common@0.0.6
+  - @biscuitland/rest@3.0.6
+
+## 3.0.5
+
+### Patch Changes
+
+- fix: logger on every module
+- Updated dependencies
+  - @biscuitland/common@0.0.5
+  - @biscuitland/rest@3.0.5
+
+## 3.0.4
+
+### Patch Changes
+
+- first 3.0.\* without bugs
+- Updated dependencies
+  - @biscuitland/common@0.0.4
+  - @biscuitland/rest@3.0.4
+
 ## 2.3.0
 
 ### Minor Changes